//Nuevos cambios 28/01/22
+Se está mejorando la implementación de omniverse en el motor.
+Se pueden guardar modelos 3D en archivos USD y exportarlo.

*Se está trabajando en poder cargar los archivos USD.
*Se está mejorando la UI.

+++ gaCore/src/gaModels.cpp
@@ -139,7 +139,7 @@ namespace gaEngineSDK {
 
   uint32
   Models::getSizeMeshes() {
-    return m_vMeshes.size();
+    return (uint32)m_vMeshes.size();
   }
 
   /***************************************************************************/
